lissajous sketch, cubic zone (work in progress)
The purpose of a Zone will be to have effect such as applying a
function only on a certain subset of vertices from a Mesh

+++ ch/supermafia/framework3D/mathematics/Function/KinectFunc.java
@@ -27,7 +27,7 @@ public class KinectFunc implements FunctionR2R3_I
 		int index = (int)(y * getUMax() + x);
 		int depth = rawDepth[index];
 		Vec3D v = depthToWorld(x, y, depth);
-		int facteur = 400;
+		float facteur = 400;
 		v.setX(facteur * v.x());
 		v.setY(facteur * v.y());
 		v.setZ(facteur - v.z() * facteur);
